requirements.txt
chardet
pycryptodome
unicodecsv>=0.14.1
pdfminer.six==20170720
pillow>=3.0.0
wand
